update lb func signature & update unit tests
