+++ src/azure-cli/azure/cli/command_modules/acs/tests/hybrid_2020_09_01/test_loadbalancer.py
@@ -13,7 +13,6 @@ from azure.cli.command_modules.acs import _loadbalancer as loadbalancer
 class TestLoadBalancer(unittest.TestCase):
 
     def test_configure_load_balancer_profile(self):
-        cmd = mock.MagicMock()
         managed_outbound_ip_count = 5
         outbound_ips = None
         outbound_ip_prefixes = None
@@ -53,18 +52,17 @@ class TestLoadBalancer(unittest.TestCase):
         }
 
         profile = ManagedClusterLoadBalancerProfile()
-        profile.managed_outbound_ips = ManagedClusterLoadBalancerProfileManagedOutboundIPs(
+        profile.managed_outbound_i_ps = ManagedClusterLoadBalancerProfileManagedOutboundIPs(
             count=2
         )
-        profile.outbound_ips = ManagedClusterLoadBalancerProfileOutboundIPs(
-            public_ips="public_ips"
+        profile.outbound_i_ps = ManagedClusterLoadBalancerProfileOutboundIPs(
+            public_i_ps="public_i_ps"
         )
         profile.outbound_ip_prefixes = ManagedClusterLoadBalancerProfileOutboundIPPrefixes(
             public_ip_prefixes="public_ip_prefixes"
         )
 
         p = loadbalancer.configure_load_balancer_profile(
-            cmd,
             managed_outbound_ip_count,
             outbound_ips,
             outbound_ip_prefixes,
@@ -74,9 +72,11 @@ class TestLoadBalancer(unittest.TestCase):
             lb_models,
         )
 
-        self.assertIsNotNone(p.managed_outbound_i_ps)
-        self.assertIsNone(p.outbound_i_ps)
-        self.assertIsNone(p.outbound_ip_prefixes)
+        self.assertEqual(p.managed_outbound_i_ps.count, 5)
+        self.assertEqual(p.outbound_i_ps, None)
+        self.assertEqual(p.outbound_ip_prefixes, None)
+        self.assertEqual(p.allocated_outbound_ports, 80)
+        self.assertEqual(p.idle_timeout_in_minutes, 3600)
 
 
 if __name__ == '__main__':
