+++ src/azure-cli/azure/cli/command_modules/acs/tests/latest/test_decorator.py
@@ -4,23 +4,18 @@
 # --------------------------------------------------------------------------------------------
 
 import importlib
-from knack import CLI
 from knack.util import CLIError
-import tempfile
 import unittest
 from unittest.mock import Mock, patch
 
-from azure.cli.core import AzCommandsLoader
 from azure.cli.core.azclierror import (
     MutuallyExclusiveArgumentError,
     RequiredArgumentMissingError,
     InvalidArgumentValueError,
 )
-from azure.cli.core.cloud import get_active_cloud
-from azure.cli.core.commands import AzCliCommand
 from azure.cli.core.profiles import ResourceType
-from azure.cli.core._config import ENV_VAR_PREFIX
 
+from azure.cli.command_modules.acs.tests.latest.mocks import MockClient, MockCLI, MockCmd
 from azure.cli.command_modules.acs._consts import (
     CONST_OUTBOUND_TYPE_USER_DEFINED_ROUTING,
 )
@@ -29,50 +24,6 @@ from azure.cli.command_modules.acs.decorator import (
     AKSCreateContext,
     AKSCreateDecorator,
 )
-
-
-MOCK_CLI_CONFIG_DIR = tempfile.mkdtemp()
-MOCK_CLI_ENV_VAR_PREFIX = "MOCK_" + ENV_VAR_PREFIX
-
-
-class MockClient(object):
-    def __init__(self):
-        pass
-
-
-class MockCLI(CLI):
-    def __init__(self):
-        super(MockCLI, self).__init__(
-            cli_name="mock_cli",
-            config_dir=MOCK_CLI_CONFIG_DIR,
-            config_env_var_prefix=MOCK_CLI_ENV_VAR_PREFIX,
-        )
-        self.cloud = get_active_cloud(self)
-
-
-class MockCmd(object):
-    def __init__(self, cli_ctx):
-        self.cli_ctx = cli_ctx
-        self.cmd = AzCliCommand(AzCommandsLoader(cli_ctx), "mock-cmd", None)
-
-    def supported_api_version(
-        self,
-        resource_type=None,
-        min_api=None,
-        max_api=None,
-        operation_group=None,
-        parameter_name=None,
-    ):
-        return self.cmd.supported_api_version(
-            resource_type=resource_type,
-            min_api=min_api,
-            max_api=max_api,
-            operation_group=operation_group,
-            parameter_name=parameter_name,
-        )
-
-    def get_models(self, *attr_args, **kwargs):
-        return self.cmd.get_models(*attr_args, **kwargs)
 
 
 class AKSCreateModelsTestCase(unittest.TestCase):
